requirements.txt
psutil>=5.9.0
PyYAML>=6.0
requests>=2.31.0
